(chore) clinical dashboard cleanups
