+++ packages/esm-cervical-cancer-app/src/views/cacx-visits/cacx-visits-services.component.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Tabs, Tab, TabList, TabPanels, TabPanel } from '@carbon/react';
 import styles from '../common.scss';
-import CaCxRegistrationList from './Tabs/cacx-registration.component';
-import CacxTreatmentList from './Tabs/cacx-treatment.component';
+import CaCxRegistrationList from './tabs/cacx-registration.component';
+import CacxTreatmentList from './tabs/cacx-treatment.component';
 import { useTranslation } from 'react-i18next';
 
 interface OverviewListProps {
