+++ packages/esm-hiv-app/src/views/hts/lab-results/lab-results-form-viewer.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react';
 import { launchFormWithCustomTitle } from '@ohri/openmrs-esm-ohri-commons-lib';
 import { getForm, applyFormIntent } from '@openmrs/openmrs-form-engine-lib';
-import styles from './Tabs/patient-list.scss';
+import styles from './tabs/patient-list.scss';
 import { OverflowMenu, OverflowMenuItem } from '@carbon/react';
 import { changeWorkspaceContext, closeAllWorkspaces, resetWorkspaceStore } from '@openmrs/esm-patient-common-lib';
 import { navigate } from '@openmrs/esm-framework';
 import { useTranslation } from 'react-i18next';
-import { moduleName } from '../../../../index';
+import { moduleName } from '../../../index';
 
 export interface LabresultsFormViewerProps {
   patientUuid: string;
