+++ packages/esm-hiv-app/src/views/hts/lab-results/tabs/patient-list.scss
@@ -1,5 +1,5 @@
 @use '@carbon/styles/scss/spacing';
-@import '../../../../../root.scss';
+@import '../../../../root.scss';
 
 .widgetContainer {
   background-color: $ui-background;
