+++ packages/esm-ohri-core-app/src/constants.ts
@@ -6,11 +6,6 @@ export const daysDurationUnit = {
 };
 
 export const basePath = '${openmrsSpaBase}/patient/';
-export const encounterRepresentation =
-  'custom:(uuid,encounterDatetime,encounterType,location:(uuid,name),' +
-  'patient:(uuid,display),encounterProviders:(uuid,provider:(uuid,name)),' +
-  'obs:(uuid,obsDatetime,voided,groupMembers,concept:(uuid,name:(uuid,name)),value:(uuid,name:(uuid,name),' +
-  'names:(uuid,conceptNameType,name))))';
 
 // Final HIV Test Result Concepts
 export const finalHIVCodeConcept = 'e16b0068-b6a2-46b7-aba9-e3be00a7b4ab';
