+++ packages/esm-ohri-core-app/src/root.scss
@@ -98,8 +98,6 @@ div[class*='-esm-login__styles__center'] > img {
   width: 15rem !important;
 }
 
-// Vitals and Biometrics Overrides
-
 div[class*='-esm-patient-vitals__vitals-header__default-border'] {
   display: flex !important;
   background-color: white !important;
